folder cleanup and readme

+++ install-dev-templates.ps1
@@ -1,6 +1,7 @@
 dotnet new uninstall Prism.Avalonia.Templates
-Remove-Item bin/**/*.nupkg
-Remove-Item obj/**/ -Force -Recurse
+
+if (Test-Path "bin/") { Remove-Item bin/**/*.nupkg }
+if (Test-Path "obj/") { Remove-Item obj/**/ -Force -Recurse }
 
 $result = dotnet pack | select-string "Successfully created package '(.*)'" -AllMatches
 $package = $result.Matches.Groups[1]
